alterado backend, adicionada versão inicial do frontend

+++ backend/CadastroLivros/Data/Entities/Autor.cs
@@ -2,7 +2,7 @@
 {
     public class Autor
     {
-        public int CodAu { get; set; }
+        public long CodAu { get; set; }
         public string Nome { get; set; }
     }
 }
