+++ backend/CadastroLivros/Services/Interfaces/IAssuntoService.cs
@@ -4,10 +4,10 @@ namespace CadastroLivros.Services.Interfaces
 {
     public interface IAssuntoService
     {
-        Task<AssuntoDTO?> Read(int codAs);
+        Task<AssuntoDTO?> Read(long codAs);
         Task<AssuntoDTO> Insert(AssuntoDTO assunto);
         Task Update(AssuntoDTO assunto);
-        Task Delete(int codAs);
+        Task Delete(long codAs);
         Task<IEnumerable<AssuntoDTO>> ReadPage(int page = 0, int pageSize = 40);
     }
 }
