+++ backend/CadastroLivros/Services/Interfaces/ILivroService.cs
@@ -4,10 +4,10 @@ namespace CadastroLivros.Services.Interfaces
 {
     public interface ILivroService
     {
-        Task<LivroDTO?> Read(int codL);
+        Task<LivroDTO?> Read(long codL);
         Task<LivroDTO> Insert(LivroDTO livro);
         Task Update(LivroDTO livro);
-        Task Delete(int codL);
+        Task Delete(long codL);
         Task<IEnumerable<LivroDTO>> ReadPage(int page = 1, int pageSize = 40);
     }
 }
